requirements.txt
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
decorator==5.1.1
future==0.18.3
geocoder==1.38.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
install==1.3.5
joblib==1.3.2
numpy==1.26.2
Pillow==10.1.0
PyAudio==0.2.14
pypiwin32==223
python-dotenv==1.0.0
pyttsx3==2.90
pywin32==306
ratelim==0.1.6
requests==2.31.0
rfc3986==1.5.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
threadpoolctl==3.2.0
urllib3==2.1.0
